Update: Refactored the product code for canvas disgonal to handle the decimal values

+++ src/calculateCanvasDiagonal/calculateCanvasDiagonal.ts
@@ -2,5 +2,5 @@ export default function calculateCanvasDiagonal(
   length: string,
   width: string
 ) {
-  return Math.sqrt(Math.pow(parseInt(length), 2) + Math.pow(parseInt(width), 2));
+  return Math.sqrt(Math.pow(parseFloat(length), 2) + Math.pow(parseFloat(width), 2));
 }
